Simplify camera properties UI to use normalized values directly
Remove conversion logic between percentage (0-100) and normalized (0.0-1.0)
values in CameraControlUI. Now uses float sliders directly in 0.0-1.0 range
to match ICameraSource interface specification.

Changes:
- Replace int property variables with float in CameraControlUI.h
- Change ImGui::SliderInt to ImGui::SliderFloat with 0.0-1.0 range
- Remove conversion code in UpdateCameraProperty and SyncUIWithCameraProperties
- Update constructor and ResetPropertiesToDefaults to use 0.5f defaults
- Simplify logging to remove redundant conversion messages

Benefits: cleaner code, better precision, API consistency, fewer calculations.

+++ src/camera/ui/CameraControlUI.h
@@ -55,10 +55,10 @@ private:
 
     // UI state
     bool m_previewEnabled;
-    int m_brightness;
-    int m_contrast;
-    int m_saturation;
-    int m_gain;
+    float m_brightness;
+    float m_contrast;
+    float m_saturation;
+    float m_gain;
 
     // Frame preview
     std::shared_ptr<const CameraFrame> m_currentFrame;
@@ -87,7 +87,7 @@ private:
     void DrawCameraInfo();
 
     // Property handling
-    void UpdateCameraProperty(CameraPropertyType property, int value);
+    void UpdateCameraProperty(CameraPropertyType property, float value);
     void ResetPropertiesToDefaults();
     void SyncUIWithCameraProperties();
